Upgrade to Rails 4.2 !! (note)
First time I've upgraded a rails app. Feels good to check this off. _phew_ This code needs so much work but this is a start.

Still have work to do but this is worth committig...

I can't believe I was using haml. I know, it's what the cool kids were using but :facepalm: